go add best practices
