+++ client/pkg/file/user_file_test.go
@@ -1,4 +1,4 @@
-package utils
+package file
 
 import (
 	"testing"
