--- client/utils/file.util_test.go
@@ -1,10 +1,0 @@
-package utils
-
-//func TestGetConfigDir(t *testing.T) {
-//	dir := getConfigDir()
-//
-//	assert.DirExists(t, dir)
-//
-//	_, err := os.Stat(dir)
-//	assert.NoError(t, err)
-//}
